Rescue split write engine writers
